Rename XMLCodeSample-ProcessingOrderForm to XMLCodeSample--ProcessingOrderForm
The subject of this project is a small company which deals with the processing and manipulation of images. This company is considering transferring some of its business processes to the Web in the form of web services. As a part of feasibility study, I was asked to carry out prototyping activities to act as proof-of-concept for this transformation.
When an image is to be processed, a paper form is filled in by an expert from the company to specify what changes the image should undergo. The following form represents a sample form filled in for an image to be processed for a client.
Processing order form 

CLIENT DETAILS: 

Organisation name: 	[J Jarndyce Ltd 	] 
Postal address: 	[Bleak House [Fort Road [Broadstairs [CT10 1EY 	] ] ] ] 
Email: 	[jjarndyce@hotmail.com 	] 
Phone: 	[01234 567890 	] 
Account number (if applicable): 	[1783 	] 
FIX REQUIRED: 
Processing instructions: 	􀂅correct exposure 
􀂅resize image 􀀻colour balance 􀂅sharpen 􀂅blur background 􀀻create composite 􀂅repair image fading 􀂅other (give details below) 
For composite: 	􀁻image attached 􀁾library image required 
Library search keywords: 	[european courtroom judge law 	] 

Special instructions: 	[resize to A4 [ [ [ 	] ] ] ] 
Date submitted: 	[20 Oct 2008 	] 
Results 
Image size: 	􀁻small (< 2 megapixels) 􀁻medium (2–6 megapixels) 􀁾large (> 6 megapixels) 
Return results by: 	􀁾CD-ROM 􀁻email 􀁻hard copy 

To demonstrate that it is feasible to use an XML document to capture image manipulation requests for the company, I was asked to create a sample XML document and an accompanying XML Schema instance. The XML document needed to contain some sample data illustrating that it could contain the same information as the request form above. The XML Schema instance was needed to validate the XML document.